rpa/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
et_xmlfile==2.0.0
idna==3.11
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
urllib3==2.6.3
